backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
httpx
python-dotenv
duckduckgo-search
ddgs
PyPDF2
python-docx
python-pptx
beautifulsoup4
lxml
numpy
scikit-learn
joblib
datasets
pytest
pytest-asyncio
pytest-xdist
tqdm
